Made all ints long long and all floats long double, now supporting numbers up to 9,223,372,036,854,775,807 :)

+++ src/utils/cpp_utils/rt_funcs/v_dequ.cpp
@@ -2,22 +2,22 @@
         value_t val_1 = val_1_v.get_value();
         value_t val_2 = value;
 
-        if (std::holds_alternative<int>(val_1) && std::holds_alternative<float>(val_2)) {
-            return std::get<int>(val_1) == std::get<float>(val_2);
+        if (std::holds_alternative<long long>(val_1) && std::holds_alternative<long double>(val_2)) {
+            return std::get<long long>(val_1) == std::get<long double>(val_2);
         }
-        else if (std::holds_alternative<float>(val_1) && std::holds_alternative<int>(val_2)) {
-            return std::get<float>(val_1) == std::get<int>(val_2);
+        else if (std::holds_alternative<long double>(val_1) && std::holds_alternative<long long>(val_2)) {
+            return std::get<long double>(val_1) == std::get<long long>(val_2);
         }
 
         if (val_1.index() != val_2.index()){
             return false;
         }
 
-        if (std::holds_alternative<int>(val_1)){
-            return std::get<int>(val_1) == std::get<int>(val_2);
+        if (std::holds_alternative<long long>(val_1)){
+            return std::get<long long>(val_1) == std::get<long long>(val_2);
         }
-        else if (std::holds_alternative<float>(val_1)){
-            return std::get<float>(val_1) == std::get<float>(val_2);
+        else if (std::holds_alternative<long double>(val_1)){
+            return std::get<long double>(val_1) == std::get<long double>(val_2);
         }
         else if (std::holds_alternative<bool>(val_1)){
             return std::get<bool>(val_1) == std::get<bool>(val_2);
