+++ src/utils/cpp_utils/rt_funcs/vadd.cpp
@@ -2,20 +2,20 @@
         value_t val_1 = val_1_v.get_value();
         value_t val_2 = val_2_v.get_value();
 
-        if (std::holds_alternative<int>(val_1) && std::holds_alternative<int>(val_2)) {
-            int raw_res = std::get<int>(val_1) + std::get<int>(val_2);
+        if (std::holds_alternative<long long>(val_1) && std::holds_alternative<long long>(val_2)) {
+            long long raw_res = std::get<long long>(val_1) + std::get<long long>(val_2);
             return Value(raw_res);
         }
-        else if (std::holds_alternative<int>(val_1) && std::holds_alternative<float>(val_2)) {
-            float raw_res = std::get<int>(val_1) + std::get<float>(val_2);
+        else if (std::holds_alternative<long long>(val_1) && std::holds_alternative<long double>(val_2)) {
+            long double raw_res = std::get<long long>(val_1) + std::get<long double>(val_2);
             return Value(raw_res);
         }
-        else if (std::holds_alternative<float>(val_1) && std::holds_alternative<int>(val_2)) {
-            float raw_res = std::get<float>(val_1) + std::get<int>(val_2);
+        else if (std::holds_alternative<long double>(val_1) && std::holds_alternative<long long>(val_2)) {
+            long double raw_res = std::get<long double>(val_1) + std::get<long long>(val_2);
             return Value(raw_res);
         }
-        else if (std::holds_alternative<float>(val_1) && std::holds_alternative<float>(val_2)) {
-            float raw_res = std::get<float>(val_1) + std::get<float>(val_2);
+        else if (std::holds_alternative<long double>(val_1) && std::holds_alternative<long double>(val_2)) {
+            long double raw_res = std::get<long double>(val_1) + std::get<long double>(val_2);
             return Value(raw_res);
         }
         else if (std::holds_alternative<std::string>(val_1) && std::holds_alternative<std::string>(val_2)) {
